in thermo_update_temps.php go back to not requiring the time to be passed in from a shell script. This method of getting the time seems to be equivalent and will base the timezone off of the one in the config.php file
In create_tables.sql.IN modify it to create the initial tables to include
the new setpoint table.  Then no need to run the setpoint update script.
This change also incorporates some other table updates included in the
setpoint update script.

We should probably start calling this "version 3" to avoid confusion of which
table updates need to be done when.